setMovement related2

+++ src/Domain/Objects/Atom.java
@@ -18,7 +18,7 @@ public class Atom extends GameObject {
         int L=GameConfiguration.getInstance().getData().getL();
         int angle = GameConfiguration.getInstance().getShooter().getAngle();
         setMovement(new MovementofObject((int) (-L / Math.tan(angle)),
-                (int) GameConfiguration.getInstance().getData().getL(), 0,true));
+                GameConfiguration.getInstance().getData().getL(), 0,true));
 
         setRectangle(new Rectangle(position, widthCoef*L, heightCoef*L, angle));
